+++ src/main/java/kr/hhplus/be/server/domain/order/OrderRepository.java
@@ -1,0 +1,7 @@
+package kr.hhplus.be.server.domain.order;
+
+public interface OrderRepository {
+
+    Order save(Order order);
+
+}
