--- src/main/java/kr/hhplus/be/server/infrastructure/order/OrderRepository.java
@@ -1,9 +1,0 @@
-package kr.hhplus.be.server.infrastructure.order;
-
-import kr.hhplus.be.server.domain.order.Order;
-
-public interface OrderRepository {
-
-    Order save(Order order);
-
-}
